pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resilient-decision-platform"
version = "0.1.0"
description = "Workflow operations platform with configurable decisioning, auditability, and retry handling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "pydantic>=2.11.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "uvicorn>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.3.0,<9.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
